loonggpu: Remove test for !screen_base in fbdev probing
The screen_base field comes from the fbdev BO and contains the fbdev
framebuffer base address. We get the BO memory via gsgpu_bo_kmap(),
which already reports the error -ENOMEM if the buffer could not be
mapped. So remove the later test for screen_base, which will never
be NULL at this point.

Following the reference implementation change.

Link: https://git.kernel.org/torvalds/c/11b6005865e4
Signed-off-by: Xi Ruoyao <xry111@xry111.site>
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>