loonggpu: drop LG_ZONE_MANAGED_PAGES
With commit 959b0886256b ("mm: constify zone related test/getter
function"), the argument to `zone_managed_pages()' is now `const'.
The commit was included as part of v6.18.

This causes conftest.sh to fail to detect whether the kernel supports
zone-managed pages, as it would no longer match the function prototype.

Since this conftest was really used to deal with pre-5.1 kernels, it is
most definitely irrelevant for us. Nullify it to fix build with 6.18.

Signed-off-by: Mingcong Bai <jeffbai@aosc.io>